Support spacy and test service
This includes changes to summary construction mostly to guarantee the
correspondence between spacy's parse and the original message.
This also includes an endpoint called /slacktest that will load a test
channel without a call to Slacker.
This also includes a fix to allow one time load of the spacy model.